fix bug with hour, simplify birthday logic

+++ lib/Birthday/Birthday.h
@@ -10,10 +10,10 @@ public:
   static void setBirthday(uint8_t month, uint8_t day);
 
 private:
-  static int _determineOffset(uint8_t month, uint8_t day);
+  static int16_t _determineOffset(uint8_t month, uint8_t day);
 
   // TODO persist this in case of power failure
-  static uint64_t _birthdays;
+  static int16_t _birthdays;
 };
 
 #endif
